Start SDK from JS only
