+++ index.js
@@ -34,16 +34,14 @@ const InstabugModule = {
    * Starts the SDK.
    * This is the main SDK method that does all the magic. This is the only
    * method that SHOULD be called.
-   * Should be called in constructor of the app registery component
+   * Should be called in constructor of the AppRegistry component
    * @param {string} token The token that identifies the app, you can find
    * it on your dashboard.
    * @param {invocationEvent} invocationEvent The event that invokes
    * the SDK's UI.
    */
   start: function (token, invocationEvent) {
-    if (Platform.OS === 'ios') {
-      Instabug.start(token, invocationEvent);
-    }
+    Instabug.start(token, invocationEvent);
     _isFirstScreen = true;
     _currentScreen = firstScreen;
     setTimeout(function () {
